Implementation of the method to safe the submited assignment

+++ Backend/src/main/java/uk/ac/rhul/cs2800/repository/AssignmentSubmissionRepository.java
@@ -4,7 +4,8 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import uk.ac.rhul.cs2800.model.AssignmentSubmission;
 
-public interface AssignmentSubmissionRepository extends JpaRepository<AssignmentSubmission, Long> {
+public interface AssignmentSubmissionRepository
+    extends JpaRepository<AssignmentSubmission, Integer> {
 
-  Optional<AssignmentSubmission> findByStudentIdAndAssignmentId(int studentId, Long assignmentId);
+  Optional<AssignmentSubmission> findByStudentIdAndAssignmentId(int studentId, int assignmentId);
 }
